[slow_ssh] Moving `generateSuggestionIp`, `conflictInterface` and `getInterfacesIps` to utils.net module.
